requirements.txt
# Core packages
xclim>=0.48.0
xarray>=2024.1.0
dask[complete]>=2024.1.0
netCDF4>=1.6.0
h5netcdf>=1.2.0

# Zarr support - primary data format
zarr>=2.16.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
click>=8.1.0
psutil>=5.9.0  # For memory monitoring

# Optional: for visualization
matplotlib>=3.7.0

# Optional: for performance monitoring
dask-jobqueue>=0.8.0
distributed>=2024.1.0